[MHV-33633] Added hideReplyButton functionality for when CannotReplyAlert is active
