+++ src/applications/mhv/secure-messaging/actions/messages.js
@@ -1,3 +1,4 @@
+import { format, addDays } from 'date-fns';
 import { Actions } from '../util/actionTypes';
 import {
   getMessageList,
@@ -75,6 +76,21 @@ export const retrieveMessage = (
       type: isDraft ? Actions.Draft.GET : Actions.Message.GET,
       response,
     });
+  }
+
+  // Error handling for old messages
+  const { sentDate } = response.data.attributes;
+  const today = new Date();
+  const messageSentDate = format(new Date(sentDate), 'MM-dd-yyyy');
+  const cannotReplyDate = addDays(new Date(messageSentDate), 45);
+  if (today > cannotReplyDate) {
+    dispatch(
+      addAlert(
+        Constants.ALERT_TYPE_INFO,
+        Constants.Alerts.Message.CANNOT_REPLY_INFO_HEADER,
+        Constants.Alerts.Message.CANNOT_REPLY_BODY,
+      ),
+    );
   }
 };
