+++ src/applications/mhv/secure-messaging/containers/MessageDetails.jsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useState } from 'react';
-import { format, addDays } from 'date-fns';
 import { useSelector, useDispatch } from 'react-redux';
 import { useLocation, useParams, useHistory } from 'react-router-dom';
 import NavigationLinks from '../components/NavigationLinks';
@@ -8,12 +7,13 @@ import { retrieveMessage } from '../actions/messages';
 import MessageDetailBlock from '../components/MessageDetailBlock';
 import AlertBackgroundBox from '../components/shared/AlertBackgroundBox';
 import AlertBox from '../components/shared/AlertBox';
-import { addAlert, closeAlert } from '../actions/alerts';
+import { closeAlert } from '../actions/alerts';
 import * as Constants from '../util/constants';
 
 const MessageDetail = () => {
   const { messageId } = useParams();
   const dispatch = useDispatch();
+  const alert = useSelector(state => state.sm.alerts.alert);
   const message = useSelector(state => state.sm.messageDetails.message);
   const messageHistory = useSelector(
     state => state.sm.messageDetails.messageHistory,
@@ -24,7 +24,8 @@ const MessageDetail = () => {
   const location = useLocation();
   const history = useHistory();
   const [id, setid] = useState(null);
-  // Pick up on adding useState for noReplyMessage
+  const [CannotReplyAlert, setCannotReplyAlert] = useState(true);
+
   useEffect(
     () => {
       if (activeFolder?.folderId === Constants.DefaultFolders.DRAFTS.id) {
@@ -39,23 +40,14 @@ const MessageDetail = () => {
     [dispatch, location, messageId, id, activeFolder, history],
   );
 
-  const isOldMessage = useEffect(
+  useEffect(
     () => {
-      const { sentDate } = messageId.data.attributes;
-      const today = new Date();
-      const messageSentDate = format(new Date(sentDate), 'MM-dd-yyyy');
-      const cannotReplyDate = addDays(new Date(messageSentDate), 45);
-      if (today > cannotReplyDate) {
-        dispatch(
-          addAlert(
-            Constants.ALERT_TYPE_INFO,
-            Constants.Alerts.Message.CANNOT_REPLY_INFO_HEADER,
-            Constants.Alerts.Message.CANNOT_REPLY_BODY,
-          ),
-        );
+      if (alert?.header !== null) {
+        setCannotReplyAlert(CannotReplyAlert);
       }
+      dispatch(closeAlert());
     },
-    [dispatch, messageId],
+    [CannotReplyAlert, alert?.header, dispatch],
   );
 
   let pageTitle;
@@ -98,9 +90,8 @@ const MessageDetail = () => {
 
   return (
     <div className="vads-l-grid-container vads-u-margin-top--2 message-detail-container">
-      {isOldMessage && <AlertBox />}
-
-      <AlertBackgroundBox closeable />
+      {/* Only display this type of alert when it contains a header */}
+      {CannotReplyAlert ? <AlertBox /> : <AlertBackgroundBox closeable />}
       <h1 className="vads-u-margin-top--2">{pageTitle}</h1>
 
       <NavigationLinks messageId={id} />
